Make conflicts between branch and database less strict. (#33)
It is an error to specify `database` and `branch` at the same
time on the same configuration level.
It is OK to override previous configuration with either `database` or
`branch`.
It is OK for the credentials file to contain both `database` and
`branch` fields in order to facilitate transitioning from old style to
the new style while maintaining backwards compatibility. The values must
agree, though.